set HealthCheckType to EC2
Having Health Check Type=ELB is dangerous because it might cause death
of the whole cluster and we already experienced such behaviour. AWS
decided to terminate nodes one by one just because of failing health
check.

In addition to that I've returned separate mount of /data to the
buku.yaml. Without it it's not possible to do major upgrade.