Calculate replication factor for the whole topic instead of partition
Couple of times we had situations when there were 4 brokers assigned for
one partition. It has happened when one of the kafka nodes was
terminated while another node was doing rebalancing. New node was not
able to start rebalancing because NotEnoughBrokersException was raised
and process of rebalancing didn't even started.
As a simple workaround of such situation we must calculate replication
factor for all partitions of one topic and choose most common value.

Example: {'0': [1,2,3], '1': [1,2,3], '2': [1,2,3,4], '3': [2,3,4]}
replication factor for all partitions must be 3, because there are 3
partitions with replication factor 3 and only one with replication
factor 4.